# lib/api-zod/src/generated/types/authSession.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UserProfile } from './userProfile';

export interface AuthSession {
  authenticated: boolean;
  vaultUnlocked: boolean;
  hasPasscode: boolean;
  user: UserProfile | null;
}