# lib/api-zod/src/generated/types/fileInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface FileInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  originalName: string;
  /** @minLength 1 */
  storagePath: string;
  mimeType: string;
  /** @minimum 0 */
  fileSize: number;
  /** @nullable */
  folderId?: string | null;
}