# lib/api-zod/src/generated/types/fileUpdate.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface FileUpdate {
  /**
     * @minLength 1
     * @nullable
     */
  name?: string | null;
  /** @nullable */
  folderId?: string | null;
}